backend/requirements.txt
fastapi[all]==0.115.12
uvicorn==0.22.0
pydantic>=2.5.2
asyncpg==0.27.0
python-dotenv==1.0.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
PyJWT==2.8.0
minio==7.1.2
SQLAlchemy==2.0.19
